+++ spec/system_helper.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require "rails_helper"
+require "base_spec_helper"
 
 # system/support/ files contain system tests configurations and helpers
 Dir[File.join(__dir__, "system/support/**/*.rb")].sort.each { |file| require file }
